Replace path filtering during publish action with individual action runs
It is likely that individual packages will be versioned individually,
rather than a common version tag across the whole monorepo. For this
reason, it may be better to run a separate build/publish step for each
package. In this case we can use default GitHub path filtering to
determine which package to build/publish.