Add licenses to individual package readmes, add formatted headers, centralise development use instructions
Users of these libraries would expect to see a `pip install` command at
the top of the readme. These have been added, as well as individual license
files for each package (so they can be consumed in PyPI). I have moved
the general "how to contribute to these libraries" instructions to the
monorepo readme via a GitHub link.